+++ src/model/OperationBancaire.java
@@ -1,31 +1,28 @@
 package model;
 
-import javafx.beans.property.IntegerProperty;
-import javafx.beans.property.SimpleIntegerProperty;
-import javafx.beans.property.SimpleStringProperty;
-import javafx.beans.property.StringProperty;
+import javafx.beans.property.*;
 
 public class OperationBancaire {
 
     private IntegerProperty id_operation_bancaire;
     private StringProperty date_operation_bancaire;
-    private IntegerProperty montant_operation_bancaire;
+    private DoubleProperty montant_operation_bancaire;
     private StringProperty libelle_type_operation;
 
 //-------------------------------------------------------------------------------------------------------------------------------------------------------------
 
-    public OperationBancaire(IntegerProperty id_operation_bancaire, StringProperty date_operation_bancaire, IntegerProperty montant_operation_bancaire, StringProperty libelle_type_operation) {
+    public OperationBancaire(IntegerProperty id_operation_bancaire, StringProperty date_operation_bancaire, DoubleProperty montant_operation_bancaire, StringProperty libelle_type_operation) {
         this.id_operation_bancaire = id_operation_bancaire;
         this.date_operation_bancaire = date_operation_bancaire;
         this.montant_operation_bancaire = montant_operation_bancaire;
         this.libelle_type_operation = libelle_type_operation;
     }
 
-    public OperationBancaire(Integer id_operation_bancaire, String date_operation_bancaire, Integer montant_operation_bancaire, String libelle_type_operation){
+    public OperationBancaire(Integer id_operation_bancaire, String date_operation_bancaire, Double montant_operation_bancaire, String libelle_type_operation){
 
         this.id_operation_bancaire = new SimpleIntegerProperty(id_operation_bancaire);
         this.date_operation_bancaire =new SimpleStringProperty(date_operation_bancaire);
-        this.montant_operation_bancaire = new SimpleIntegerProperty(montant_operation_bancaire);
+        this.montant_operation_bancaire = new SimpleDoubleProperty(montant_operation_bancaire);
         this.libelle_type_operation = new SimpleStringProperty(libelle_type_operation);
     }
 
@@ -51,15 +48,15 @@ public class OperationBancaire {
         this.date_operation_bancaire = date_operation_bancaire;
     }
 
-    public int getMontant_operation_bancaire() {
+    public Double getMontant_operation_bancaire() {
         return montant_operation_bancaire.get();
     }
 
-    public IntegerProperty montant_operation_bancaireProperty() {
+    public DoubleProperty montant_operation_bancaireProperty() {
         return montant_operation_bancaire;
     }
 
-    public void setMontant_operation_bancaire(int montant_operation_bancaire) {
+    public void setMontant_operation_bancaire(Double montant_operation_bancaire) {
         this.montant_operation_bancaire.set(montant_operation_bancaire);
     }
